Update input params
